VendorRisk entity moved to domain.layer
